third_party/WebKit/LayoutTests/accessibility/aria-disabled-expected.txt

 
link submit
treeitem
option
tab
gridcell
checkbox

This tests that the aria-disabled attribute works. The text field should not be enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textField.restriction is "disabled"
PASS textField.restriction is "none"
PASS group.restriction is "disabled"
PASS link.restriction is "disabled"
PASS button.restriction is "none"
PASS treeitem.restriction is "disabled"
PASS option.restriction is "disabled"
PASS tab.restriction is "disabled"
PASS gridcell.restriction is "disabled"
PASS checkbox.restriction is "disabled"
PASS successfullyParsed is true

TEST COMPLETE
